Finalização

+++ Aula Aventureiro1/HoradeCodar.c
@@ -3,7 +3,7 @@
 int main(){
 
     int nota1, nota2, nota3;
-    int media;
+    float media;
 
     printf("*** PROGRAMA DE CÁLCULO DE MÉDIA *** \n");
 
@@ -16,9 +16,9 @@ int main(){
     printf("Digite sua terceira nota: \n");
     scanf("%d", &nota3);
 
-    media = (nota1 + nota2 + nota3) / 3;
+    media = (float)(nota1 + nota2 + nota3) / 3;
 
-    printf("A média é: %d \n", media);
+    printf("A média é: %.1f \n", media);
